FourSidedCavityFlow.CavityCache — Type

A struct containing the cache variables for the nonlinear function f.

FourSidedCavityFlow.CavityParameters — Type

A struct containing the parameters for the four-sided cavity flow. The fixed parameters consist of the fields for the discretization and the boundary conditions. The field Re is meant to be set beforehand to run a simulation for a given Reynolds number.

FourSidedCavityFlow.CavityStruct — Type

The CavityStruct contains the sub-structs CavityParameters and CavityCache. It is passed to to the function f!(fu, u, p::CavityStruct) to provide the parameters and the cache for the evaluation of the nonlinear function.

FourSidedCavityFlow.constructBC! — Method

construct!BC(Ψ, p)

Constructs the two outer grid points rows and columns from the given boundary conditions. The outer rows and columns of the Ψ matrix will be overridden.

FourSidedCavityFlow.constructBC — Method

Ψ = constructBC(u, p)

Out-of-place version of the function constructBC!(Ψ, p). The input is the vector of inner grid points u and the function returns the Ψ matrix with the boundaries reconstructed for the given flow field.

FourSidedCavityFlow.constructBC_matrix — Method

Minv = constructBC_matrix(D)

Helper function to return the 2 by 2 matrices needed for the construction of the boundary for the streamfunction. D is the Chebyshev differentiation matrix of the discretization.

FourSidedCavityFlow.construct_homogenousBC! — Method

construct_homogenousBC(Ψ, p)

Constructs the boundaries for the streamfunction matrix Ψ by using homogenous boundary conditions (zero velocities at all lids of the cavity). This is employed in the perturbed solution for the linear stability analysis.

FourSidedCavityFlow.continuation_arclength — Method

continuation_arclength(foldercont, Ψ1, Ψ2, p::CavityStruct, Re1, Re2, Re_steps; lsa = false)

Arc-length continuation algorithm, which starts with two initial states. It is possible to turn on linear stability analysis at every continuation step. The Ψ states are saved in a binary format in the destination folder. A file called results.csv gives an overview of the computed continuation steps.

Arguments

foldercont::String: Destination folder for the continuation

Ψ1::Matrix: First streamfunction solution for the streamfunction to start continuation

Ψ2::Matrix: Second streamfunction solution for the streamfunction to start continuation

Re1::Real: Reynolds number of first solution

Re2::Real: Reynolds number of second solution

Re_step::Integer: Number of continuation steps

timesteps:Integer: Number of timesteps

lsa::Bool=false: If true a linear stability at every continuation step is performed

FourSidedCavityFlow.continuation_arclength — Method

continuation_arclength(foldercont, Ψi, p, Re_start, ΔRe, Re_steps; lsa = false)

Arc-length continuation algorithm, which starts with one initial states. The first continuation is performed by doing a natural continuation step. The Ψ states are saved in a binary format in the destination folder. A file called results.csv gives an overview of the computed continuation steps.

Arguments

foldercont::String: Destination folder for the continuation

Ψi::Matrix: Streamfunction solution for the streamfunction to start continuation

Re_start::Real: Reynolds number of start solution

ΔRe::Real: Inital step in Reynolds

Re_step::Integer: Number of continuation steps

lsa::Bool=false: If true a linear stabilility at every continuaiton step is performed

FourSidedCavityFlow.diff_chebyshev — Method

diff_chebyshev(n)

Compute the Chebyshev differentiation matrix with n + 1 points in the [1, -1]. Returns a vector of nodes and the matrices for the first, second and fourth derivatives.

FourSidedCavityFlow.f! — Method

f!(fu, u, p)

The nonlinear function for steady-state solutions of the four-sided cavity flow. Evaluates the equation:

$F(\Psi) = \frac{1}{\mathrm{Re}} \Delta^2 \Psi + (\partial_x \Psi) \partial_y(\Delta \Psi) - (\partial_y \Psi) \partial_x(\Delta \Psi)$

Only the interior part u (given as a vector) of the streamfunction matrix Ψ is needed, as the boundary conditions explicitly give the 2 outer rows and columns. fu is a vector where the solution of the function evaluation will be stored.

Example

julia> n = 64; Re = 100; 
julia> p = setup_struct(n, Re) 
julia> Ψ = zeros(n + 1, n+ 1)
julia> u = Ψ[3:(n - 1), 3:(n - 1)][:]
julia> fu = similiar(u) 
julia> f!(fu, u, p) 

FourSidedCavityFlow.f_linearstability! — Method

f_linearstability!(fu, Ψ, Ψ0, p)

Computes the for a given solution Ψ0 and the eigenfunction Ψ. This function calculates the elements of the matrix $A$ generalized eigenvalue problem. Stores the solution in fu.

FourSidedCavityFlow.ftime! — Method

ftime!(fu, u, p, h)

The nonlinear function for the time integration of the streamfunction formulation. Solves the equations discretized with an implicit Euler scheme and a timestep h:

$F(\Psi_{i+1}) = \Delta\Psi_i - \Delta\Psi_{i+1} + h (\frac{1}{\mathrm{Re}}\Delta^2 \Psi_{i+1} + (\partial_x \Psi_{i+1}) \partial_y(\Delta \Psi_{i+1}) - (\partial_y \Psi_{i+1}) \partial_x(\Delta \Psi_{i+1})) = 0$

The solution of the previous timestep has to be set beforehand by p.params.Ψi.

FourSidedCavityFlow.linearstability_lambdamax — Method

linearstability_lambdamax(Re, u, p)

Computes the real part of the rightmost eigenvalues of the linear stability around a base state u and a Reynolds number Re.

FourSidedCavityFlow.linearstability_lambdas — Method

linearstability_lambdas(u, p)

Computes the eigenvalues of the linear stability around a base state u.

FourSidedCavityFlow.linearstability_matrices! — Method

f_linearstability!(fu, Ψ, Ψ0, p)

Computes the matrices A and B of the linear stability around a base state u.

FourSidedCavityFlow.newton — Method

x, iter, tol = newton(f!, x0, p, newton_cache; abstol = 1e-10, maxiters = 100)

Newton-Raphson to compute the zero of a multiple-dimensional function f. An in-place version for the function has to be provided f(fx, x, p).

Arguments

f!::Function: In-place formulation of multi-dimensional unction, f!(fx, x, p)

x0::Real: Inital guess

p: Parameters and cache for the function

newton_cache::Real: Tuple of variables used to cache for the Newton newton_cache = (x, fx, dx, J, jac_cache), jac_cache is created for the Jacobian calculation through the package FiniteDiff.jl

abstol::Real=1e-10: Absolut tolerance

maxiters::Integer=100: Maximum iterations for the Newton algorithm

FourSidedCavityFlow.newton — Method

x, iter, tol = newton(f!, x0, p; abstol = 1e-10, maxiters = 100)

Newton-Raphson to compute the zero of a multiple-dimensional function f. An in-place version for the function has to be provided f(fx, x, p). This function version doesn't use a cache for the Newton and allocates at every call.

Arguments

f!::Function: In-place formulation of multi-dimensional function, f!(fx, x, p)

x0::Real: Inital guess

p: Parameters and cache for the function

abstol::Real=1e-10: Absolut tolerance

maxiters::Integer=100: Maximum iterations for the Newton algorithm

FourSidedCavityFlow.newton1D — Function

x, iter, tol = newton1D(f::Function, x0, p; abstol = 1e-10, maxiters = 100)

1D Newton method. Returns the computed zero x, the number of iterations iter and the absolut tolerance tol. p contains the the parameters for the function such that fx = f(x, p)

FourSidedCavityFlow.newton1D_for_linearstability — Method

Re, u, iter, tol = newton1D_for_linearstability(Re0, u0, p; abstol = 1e-10, maxiters = 20, verbose = false)

1D Newton to calculates the Reynolds number `Re` where the rightmost eigenvalues real part is 0. `u0` is the initial guess where
for the bifurcation. It only works if the state where the bifurcation is has a well-conditioned Jacobian.

FourSidedCavityFlow.setup_struct — Method

setup_struct(n, Re)

Function to instantiate the CavityStruct. Creates a grid of size $(n+1) \times (n+1)$. The regularization parameter for the boundary conditions is fixed to $k_0 = 10$.

Example

julia> p = setup_struct(64, 1)

FourSidedCavityFlow.steadystate — Method

Ψsteady, iter, tol = steady_state(Ψ0, p; abstol = 1e-10, maxiters = 100)

Computes the steady-state solution for the four-sided cavity by employing the Newton algorithm. Returns the calculated steady-state solution from a given initial guess Ψ0 and the number of iterions iter and tolerance tol of Newton's method.

FourSidedCavityFlow.timestepping — Method

Ψi = timestepping(Ψinit, p, h, timesteps; savesteps = false, verbose = false)

Timestepping of the equation of motion starting with an initial solution Ψinit and then perfoms a specified number of timesteps using the implicit Euler scheme.

Arguments

Ψinit::Matrix: Initial streamfunction field to start time-integration

p::CavityStruct: Parameters and cache for the problem.

h::Real: Timestep size

timesteps:Integer: Number of timesteps

savesteps::Bool=false: If false only returns the solution of the last timesteps, if true a solution vector of all timesteps and the timeseries

verbose::Bool=false: If true prints information of the state at each timesteps